+++ uninstall.php
@@ -1,5 +1,0 @@
-<?php
-if (!defined('WP_UNINSTALL_PLUGIN')) {
-    die;
-}
-include_once('php/gb-fc-uninstall.php');